// src/app/shared/interfaces/symptoms.ts
export class Symptoms {
    fever: boolean;
    feverDeg?: number;
    cough: boolean;
    muscularPain: boolean;
    soreThroat: boolean;
    diarrhea: boolean;
    tiredness: boolean;
    dyspnea: boolean; // difficulty of breathing
    anorexia: boolean; // fo9dan chahia
    discomfort: boolean;
}